@@ -9,7 +9,7 @@ class Change extends Model
 {
     use HasFactory;
 
-    protected $dates = [
+    protected $casts = [
         'created_at',
         'updated_at',
     ];
