@@ -5,8 +5,8 @@ namespace App\Providers;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Event;
-use \SocialiteProviders\Manager\SocialiteWasCalled;
+/* use Illuminate\Support\Facades\Event; */
+/* use \SocialiteProviders\Manager\SocialiteWasCalled; */
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -20,9 +20,9 @@ class EventServiceProvider extends ServiceProvider
             SendEmailVerificationNotification::class,
         ],
 
-        SocialiteWasCalled::class => [
-            'Uspdev\SenhaunicaSocialite\SenhaunicaExtendSocialite@handle',
-        ],
+        /* SocialiteWasCalled::class => [ */
+        /*     'Uspdev\SenhaunicaSocialite\SenhaunicaExtendSocialite@handle', */
+        /* ], */
     ];
 
     /**
